move span and consolidate slice construction
